+++ tabletest-junit/src/test/java/io/github/nchaugen/tabletest/renderer/AsciidocRendererTest.java
@@ -23,23 +23,35 @@ public class AsciidocRendererTest {
     private final AsciidocRenderer renderer = new AsciidocRenderer(new StubExtensionContext());
 
     @Test
-    void shouldRenderBasicTable() {
+    void shouldAddRoleForExpectationCells() {
         assertEquals(
             """
-                [%header,cols="1,1,1"]
+                [%header,cols="1,1,1,1,1"]
                 |===
-                |a
-                |b
-                |c
+                |[.expectation]#+a?+#
+                |[.expectation]#+b?+#
+                |[.expectation]#+c?+#
+                |[.expectation]#+d?+#
+                |[.expectation]#+e?+#
                 
-                a|1
-                a|2
-                a|3
+                a|[.expectation]#{empty}#
+                a|[.expectation]
+                . +1+
+                . +2+
+                . +3+
+                
+                a|[.expectation]#+3+#
+                a|[.expectation]
+                a|[.expectation]
+                +a+:: +1+
+                +b+:: +2+
+                +c+:: +3+
+
                 |===
                 """,
             renderer.render(TableParser.parse("""
-                a | b | c
-                1 | 2 | 3
+                a? | b?      | c? | d? | e?
+                {} | [1,2,3] | 3  |    | [a:1,b:2,c:3]
                 """))
         );
     }
@@ -50,13 +62,13 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1,1,1"]
                 |===
-                |a
-                |b
-                |c
+                |+a+
+                |+b+
+                |+c+
                 
-                a|1
+                a|+1+
                 a|
-                a|3
+                a|+3+
                 |===
                 """,
             renderer.render(TableParser.parse("""
@@ -72,13 +84,13 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1,1,1"]
                 |===
-                |a
-                |b
-                |a\\|b
+                |+a+
+                |+b+
+                |+a\\|b+
                 
-                a|\\|
-                a|\\|
-                a|Text with \\| character
+                a|+\\|+
+                a|+\\|+
+                a|+Text with \\| character+
                 |===
                 """,
             renderer.render(TableParser.parse("""
@@ -94,19 +106,19 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1,1,1"]
                 |===
-                |a
-                |b
-                |c
+                |+a+
+                |+b+
+                |+c+
                 
                 a|{empty}
                 a|
-                . 1
-                . 2
-                . 3
+                . +1+
+                . +2+
+                . +3+
                 
                 a|
-                . \\|
-                . \\|
+                . +\\|+
+                . +\\|+
                 
                 |===
                 """,
@@ -123,9 +135,9 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1,1,1"]
                 |===
-                |a
-                |b
-                |c
+                |+a+
+                |+b+
+                |+c+
                 
                 a|{empty}
                 a|
@@ -150,21 +162,21 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1"]
                 |===
-                |a
+                |+a+
                 
                 a|
                 . {empty}
-                  .. 1
-                  .. 2
-                  .. 3
+                  .. +1+
+                  .. +2+
+                  .. +3+
                 . {empty}
-                  .. a
-                  .. b
-                  .. c
+                  .. +a+
+                  .. +b+
+                  .. +c+
                 . {empty}
-                  .. #
-                  .. $
-                  .. %
+                  .. +#+
+                  .. +$+
+                  .. +%+
                 
                 |===
                 """,
@@ -185,19 +197,19 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1,1,1"]
                 |===
-                |a
-                |b
-                |c
+                |+a+
+                |+b+
+                |+c+
                 
                 a|{empty}
                 a|
-                * 1
-                * 2
-                * 3
+                * +1+
+                * +2+
+                * +3+
                 
                 a|
-                * \\|
-                * \\|
+                * +\\|+
+                * +\\|+
                 
                 |===
                 """,
@@ -218,19 +230,19 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1,1,1"]
                 |===
-                |a
-                |b
-                |c
+                |+a+
+                |+b+
+                |+c+
                 
                 a|{empty}
                 a|
-                . 1
-                . 2
-                . 3
+                . +1+
+                . +2+
+                . +3+
                 
                 a|
-                . \\|
-                . \\|
+                . +\\|+
+                . +\\|+
                 
                 |===
                 """,
@@ -254,25 +266,25 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1"]
                 |===
-                |a
+                |+a+
                 
                 a|
                 [square]
                 * {empty}
                 [none]
-                  ** 1
-                  ** 2
-                  ** 3
+                  ** +1+
+                  ** +2+
+                  ** +3+
                 * {empty}
                 [none]
-                  ** a
-                  ** b
-                  ** c
+                  ** +a+
+                  ** +b+
+                  ** +c+
                 * {empty}
                 [none]
-                  ** #
-                  ** $
-                  ** %
+                  ** +#+
+                  ** +$+
+                  ** +%+
                 
                 |===
                 """,
@@ -289,18 +301,18 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1,1,1"]
                 |===
-                |a
-                |b
-                |c
+                |+a+
+                |+b+
+                |+c+
                 
                 a|{empty}
                 a|
-                * 1
-                * 2
-                * 3
+                * +1+
+                * +2+
+                * +3+
                 
                 a|
-                * \\|\\|
+                * +\\|\\|+
                 
                 |===
                 """,
@@ -317,21 +329,21 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1"]
                 |===
-                |a
+                |+a+
                 
                 a|
                 * {empty}
-                  ** 1
-                  ** 2
-                  ** 3
+                  ** +1+
+                  ** +2+
+                  ** +3+
                 * {empty}
-                  ** a
-                  ** b
-                  ** c
+                  ** +a+
+                  ** +b+
+                  ** +c+
                 * {empty}
-                  ** #
-                  ** $
-                  ** %
+                  ** +#+
+                  ** +$+
+                  ** +%+
                 
                 |===
                 """,
@@ -355,17 +367,17 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1,1"]
                 |===
-                |a
-                |b
+                |+a+
+                |+b+
                 
                 a|{empty}
                 a|
                 [lowergreek]
                 . {empty}
-                  .. 1
+                  .. +1+
                 . {empty}
-                  .. 2
-                  .. 3
+                  .. +2+
+                  .. +3+
                 
                 |===
                 """,
@@ -389,25 +401,25 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1"]
                 |===
-                |a
+                |+a+
                 
                 a|
                 [square]
                 * {empty}
                 [none]
-                  ** 1
-                  ** 2
-                  ** 3
+                  ** +1+
+                  ** +2+
+                  ** +3+
                 * {empty}
                 [none]
-                  ** a
-                  ** b
-                  ** c
+                  ** +a+
+                  ** +b+
+                  ** +c+
                 * {empty}
                 [none]
-                  ** #
-                  ** $
-                  ** %
+                  ** +#+
+                  ** +$+
+                  ** +%+
                 
                 |===
                 """,
@@ -424,18 +436,18 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1,1,1"]
                 |===
-                |a
-                |b
-                |c
+                |+a+
+                |+b+
+                |+c+
                 
                 a|{empty}
                 a|
-                a:: 1
-                b:: 2
-                c:: 3
+                +a+:: +1+
+                +b+:: +2+
+                +c+:: +3+
                 
                 a|
-                b:: \\|\\|
+                +b+:: +\\|\\|+
                 
                 |===
                 """,
@@ -452,18 +464,18 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1,1"]
                 |===
-                |a
-                |b
+                |+a+
+                |+b+
                 
                 a|
-                a:: {empty}
-                b:: {empty}
+                +a+:: {empty}
+                +b+:: {empty}
                 
                 a|
-                a::
-                  A::: 1
-                b::
-                  B::: 2
+                +a+::
+                  +A+::: +1+
+                +b+::
+                  +B+::: +2+
                 
                 |===
                 """,
@@ -480,23 +492,23 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1,1"]
                 |===
-                |a
-                |b
+                |+a+
+                |+b+
                 
                 a|
-                a::
-                  .. 1
-                  .. 2
-                b::
-                  ** 3
-                  ** 4
-                c:: 5
+                +a+::
+                  .. +1+
+                  .. +2+
+                +b+::
+                  ** +3+
+                  ** +4+
+                +c+:: +5+
                 
                 a|
                 * {empty}
-                  A::: 1
+                  +A+::: +1+
                 * {empty}
-                  B::: 2
+                  +B+::: +2+
                 
                 |===
                 """,
@@ -522,25 +534,25 @@ public class AsciidocRendererTest {
             """
                 [%header,cols="1,1"]
                 |===
-                |a
-                |b
+                |+a+
+                |+b+
                 
                 a|
-                a::
-                  ** 1
-                  ** 2
-                b::
+                +a+::
+                  ** +1+
+                  ** +2+
+                +b+::
                 [disc]
-                  ** 3
-                  ** 4
-                c:: 5
+                  ** +3+
+                  ** +4+
+                +c+:: +5+
                 
                 a|
                 [circle]
                 * {empty}
-                  A::: 1
+                  +A+::: +1+
                 * {empty}
-                  B::: 2
+                  +B+::: +2+
                 
                 |===
                 """,